mdm-writer/pyproject.toml
[project]
name = "mdm-writer"
version = "0.1.0"
description = "Writes and upserts MDM customer and product master data into MySQL"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python==9.0.0",
]

[project.scripts]
mdm-writer = "app.main:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
